Split dewy.go and container/runtime.go into thematic files
Both files had grown to >1000 lines, mixing many distinct concerns
(deploy mechanics, TCP proxy, admin HTTP API, hook execution, image
management, container inspection, credentials handling) in single
package files. This commit is a behaviour-preserving relocation: every
function and type keeps the same name, the same signature, and the
same call sites; only the file it lives in changes.

dewy package (1223 -> 313 lines for dewy.go core, +7 new files):
- dewy.go             : Dewy struct, New, SetTelemetry, Start, waitSigs,
                        cachekeyName, Run, RunContainer
- release.go          : deploy, preserve, restartServer, startServer,
                        keepReleases, cleanupOldImages
- hooks.go            : execHook
- container_deploy.go : deployContainer, createHealthCheckFunc,
                        stopManagedContainers
- proxy.go            : tcpProxy / tcpBackend types and methods,
                        startProxy, stopProxy, totalProxyBackends
- proxy_backend.go    : proxyBackendUpdater, addProxyBackend,
                        removeProxyBackend
- admin_api.go        : startAdminAPI, stopAdminAPI,
                        handleGetContainers, handleGetStatus
- connio.go           : limitedWriter, timeoutConn

container package (runtime.go 1181 -> 222 lines, +4 new files):
- runtime.go             : Runtime struct, newCLIRuntime, validateExtraArgs,
                           hasUserOption, extractNameOption,
                           execCommand, execCommandOutput, Run, Stop, Remove
- runtime_credentials.go : extractRegistry, getCredentials, isAuthError, Login
- runtime_image.go       : Pull, pullImage, ListImages, RemoveImage,
                           CleanupOldImages, GetImageExposedPorts,
                           imageRepositoryFromRef, imageInspect
- runtime_deploy.go      : Deploy, startAndCheck, rollback,
                           ResolvePortMappings, StopManagedContainers
- runtime_inspect.go     : FindContainerByLabel, FindContainersByLabel,
                           GetMappedPort, GetRunningContainerWithImage,
                           GetContainerInfo, ListContainersByLabels,
                           containerInspect

Imports were re-derived per file; nothing else changed.
