+++ container/runtime_inspect.go
@@ -1,0 +1,235 @@
+package container
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"log/slog"
+	"strings"
+	"time"
+)
+
+// containerInspect represents the structure of container inspect JSON output.
+type containerInspect struct {
+	ID      string `json:"Id"`
+	Name    string `json:"Name"`
+	Created string `json:"Created"`
+	State   struct {
+		Status    string `json:"Status"`
+		StartedAt string `json:"StartedAt"`
+	} `json:"State"`
+	Config struct {
+		Image        string              `json:"Image"`
+		Labels       map[string]string   `json:"Labels"`
+		ExposedPorts map[string]struct{} `json:"ExposedPorts"`
+	} `json:"Config"`
+	NetworkSettings struct {
+		Ports map[string][]struct {
+			HostIP   string `json:"HostIp"`
+			HostPort string `json:"HostPort"`
+		} `json:"Ports"`
+	} `json:"NetworkSettings"`
+}
+
+// FindContainerByLabel finds a container by labels.
+func (r *Runtime) FindContainerByLabel(ctx context.Context, labels map[string]string) (string, error) {
+	args := []string{"ps", "-q"}
+
+	for key, value := range labels {
+		args = append(args, "--filter", fmt.Sprintf("label=%s=%s", key, value))
+	}
+
+	output, err := r.execCommandOutput(ctx, args...)
+	if err != nil {
+		return "", err
+	}
+
+	if output == "" {
+		return "", ErrContainerNotFound
+	}
+
+	lines := strings.Split(output, "\n")
+	return lines[0], nil
+}
+
+// FindContainersByLabel finds all containers matching the given labels.
+func (r *Runtime) FindContainersByLabel(ctx context.Context, labels map[string]string) ([]string, error) {
+	args := []string{"ps", "-q"}
+
+	for key, value := range labels {
+		args = append(args, "--filter", fmt.Sprintf("label=%s=%s", key, value))
+	}
+
+	output, err := r.execCommandOutput(ctx, args...)
+	if err != nil {
+		return nil, err
+	}
+
+	if output == "" {
+		return []string{}, nil
+	}
+
+	lines := strings.Split(output, "\n")
+	containers := make([]string, 0, len(lines))
+	for _, line := range lines {
+		line = strings.TrimSpace(line)
+		if line != "" {
+			containers = append(containers, line)
+		}
+	}
+
+	r.logger.Debug("Found containers by label",
+		slog.Any("labels", labels),
+		slog.Int("count", len(containers)))
+
+	return containers, nil
+}
+
+// GetMappedPort returns the host port mapped to the container port.
+func (r *Runtime) GetMappedPort(ctx context.Context, containerID string, containerPort int) (int, error) {
+	portSpec := fmt.Sprintf("%d/tcp", containerPort)
+	output, err := r.execCommandOutput(ctx, "port", containerID, portSpec)
+	if err != nil {
+		return 0, fmt.Errorf("failed to get mapped port: %w", err)
+	}
+
+	if output == "" {
+		return 0, fmt.Errorf("no port mapping found for container port %d", containerPort)
+	}
+
+	// Output format: "0.0.0.0:32768" or "0.0.0.0:32768\n:::32768"
+	parts := strings.Split(output, "\n")
+	if len(parts) == 0 {
+		return 0, fmt.Errorf("unexpected port output format: %s", output)
+	}
+
+	firstMapping := strings.TrimSpace(parts[0])
+	colonIdx := strings.LastIndex(firstMapping, ":")
+	if colonIdx == -1 {
+		return 0, fmt.Errorf("unexpected port format: %s", firstMapping)
+	}
+
+	portStr := firstMapping[colonIdx+1:]
+	port, err := fmt.Sscanf(portStr, "%d", new(int))
+	if err != nil || port != 1 {
+		return 0, fmt.Errorf("failed to parse port number from %s: %w", portStr, err)
+	}
+
+	var hostPort int
+	n, err := fmt.Sscanf(portStr, "%d", &hostPort)
+	if err != nil {
+		return 0, fmt.Errorf("fmt.Sscanf failed: n=%d, %w", n, err)
+	}
+
+	return hostPort, nil
+}
+
+// GetRunningContainerWithImage checks if a container is running with the specified image and app name.
+func (r *Runtime) GetRunningContainerWithImage(ctx context.Context, imageRef, appName string) (string, error) {
+	output, err := r.execCommandOutput(ctx, "ps",
+		"--filter", fmt.Sprintf("ancestor=%s", imageRef),
+		"--filter", fmt.Sprintf("label=dewy.app=%s", appName),
+		"--filter", "status=running",
+		"--format", "{{.ID}}")
+
+	if err != nil {
+		return "", fmt.Errorf("failed to list running containers: %w", err)
+	}
+
+	if output == "" {
+		r.logger.Debug("No running container found with image", slog.String("image", imageRef))
+		return "", nil
+	}
+
+	containerIDs := strings.Split(output, "\n")
+	containerID := strings.TrimSpace(containerIDs[0])
+
+	r.logger.Debug("Found running container with image",
+		slog.String("image", imageRef),
+		slog.String("container", containerID))
+	return containerID, nil
+}
+
+// GetContainerInfo returns detailed information about a container.
+func (r *Runtime) GetContainerInfo(ctx context.Context, containerID string, containerPort int) (*Info, error) {
+	output, err := r.execCommandOutput(ctx, "inspect", containerID)
+	if err != nil {
+		return nil, fmt.Errorf("failed to inspect container: %w", err)
+	}
+
+	var inspects []containerInspect
+	if err := json.Unmarshal([]byte(output), &inspects); err != nil {
+		return nil, fmt.Errorf("failed to parse inspect output: %w", err)
+	}
+
+	if len(inspects) == 0 {
+		return nil, ErrContainerNotFound
+	}
+
+	inspect := inspects[0]
+
+	startedAt, err := time.Parse(time.RFC3339Nano, inspect.State.StartedAt)
+	if err != nil {
+		r.logger.Warn("Failed to parse StartedAt timestamp",
+			slog.String("container", containerID),
+			slog.String("timestamp", inspect.State.StartedAt))
+		startedAt = time.Time{}
+	}
+
+	deployedAt := startedAt
+	if deployedAtStr, ok := inspect.Config.Labels["dewy.deployed_at"]; ok {
+		if t, err := time.Parse(time.RFC3339, deployedAtStr); err == nil {
+			deployedAt = t
+		}
+	}
+
+	ipPort := ""
+	if containerPort > 0 {
+		portSpec := fmt.Sprintf("%d/tcp", containerPort)
+		if portBindings, ok := inspect.NetworkSettings.Ports[portSpec]; ok && len(portBindings) > 0 {
+			ipPort = fmt.Sprintf("%s:%s", portBindings[0].HostIP, portBindings[0].HostPort)
+		}
+	} else {
+		for _, portBindings := range inspect.NetworkSettings.Ports {
+			if len(portBindings) > 0 {
+				ipPort = fmt.Sprintf("%s:%s", portBindings[0].HostIP, portBindings[0].HostPort)
+				break
+			}
+		}
+	}
+
+	name := strings.TrimPrefix(inspect.Name, "/")
+
+	return &Info{
+		ID:         inspect.ID,
+		Name:       name,
+		Image:      inspect.Config.Image,
+		Status:     inspect.State.Status,
+		IPPort:     ipPort,
+		StartedAt:  startedAt,
+		DeployedAt: deployedAt,
+		Labels:     inspect.Config.Labels,
+	}, nil
+}
+
+// ListContainersByLabels returns detailed information about containers matching the given labels.
+func (r *Runtime) ListContainersByLabels(ctx context.Context, labels map[string]string, containerPort int) ([]*Info, error) {
+	containerIDs, err := r.FindContainersByLabel(ctx, labels)
+	if err != nil {
+		return nil, fmt.Errorf("failed to find containers: %w", err)
+	}
+
+	infos := make([]*Info, 0, len(containerIDs))
+	for _, containerID := range containerIDs {
+		info, err := r.GetContainerInfo(ctx, containerID, containerPort)
+		if err != nil {
+			r.logger.Warn("Failed to get container info",
+				slog.String("container", containerID),
+				slog.String("error", err.Error()))
+			continue
+		}
+		infos = append(infos, info)
+	}
+
+	return infos, nil
+}
